@@ -11,7 +11,8 @@
 		"axios": "^0.27.2",
 		"bootstrap": "^5.1.3",
 		"core-js": "^3.8.3",
-		"vue": "^2.6.14"
+		"vue": "^3.2.13",
+		"vue-router": "^4.0.15"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.12.16",
@@ -20,8 +21,7 @@
 		"@vue/cli-plugin-eslint": "~5.0.0",
 		"@vue/cli-service": "~5.0.0",
 		"eslint": "^7.32.0",
-		"eslint-plugin-vue": "^8.0.3",
-		"vue-template-compiler": "^2.6.14"
+		"eslint-plugin-vue": "^8.0.3"
 	},
 	"eslintConfig": {
 		"root": true,
@@ -29,7 +29,7 @@
 			"node": true
 		},
 		"extends": [
-			"plugin:vue/essential",
+			"plugin:vue/vue3-essential",
 			"eslint:recommended"
 		],
 		"parserOptions": {
@@ -42,6 +42,7 @@
 	"browserslist": [
 		"> 1%",
 		"last 2 versions",
-		"not dead"
+		"not dead",
+		"not ie 11"
 	]
 }
