@@ -1,13 +1,11 @@
-import Vue from 'vue';
+// Vue.prototype.$http = axios;
+// Vue.prototype.$axios = axios;
+// Vue.config.productionTip = false;
+
+import { createApp } from 'vue';
 import App from './App.vue';
-import axios from 'axios';
 import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
+import { router } from './assets/router';
 
-Vue.prototype.$http = axios;
-Vue.prototype.$axios = axios;
-Vue.config.productionTip = false;
-
-new Vue({
-	render: (h) => h(App),
-}).$mount('#app');
+createApp(App).use(router).mount('#app');
